NCFE-62: Reuse create file func
